ST-1130 PMD error fix
