+++ src/main/java/uk/gov/hmcts/sptribs/caseworker/util/DecisionDocumentListUtil.java
@@ -1,6 +1,7 @@
 package uk.gov.hmcts.sptribs.caseworker.util;
 
 import org.springframework.util.ObjectUtils;
+import uk.gov.hmcts.ccd.sdk.type.ListValue;
 import uk.gov.hmcts.sptribs.ciccase.model.CaseData;
 import uk.gov.hmcts.sptribs.document.model.CaseworkerCICDocument;
 import uk.gov.hmcts.sptribs.document.model.DocumentType;
@@ -8,12 +9,14 @@ import uk.gov.hmcts.sptribs.document.model.DocumentType;
 import java.util.ArrayList;
 import java.util.List;
 
+import static uk.gov.hmcts.sptribs.caseworker.util.DocumentManagementUtil.EMPTY_DOCUMENT;
+import static uk.gov.hmcts.sptribs.caseworker.util.DocumentManagementUtil.addToRemovedDocuments;
+
 public  final class DecisionDocumentListUtil {
 
     private DecisionDocumentListUtil() {
 
     }
-
 
     public static List<CaseworkerCICDocument> getDecisionDocs(CaseData caseData) {
         List<CaseworkerCICDocument> decisionDocs = new ArrayList<>();
@@ -57,6 +60,47 @@ public  final class DecisionDocumentListUtil {
             }
         }
         return finalDecisionDocs;
+    }
+
+    public static void removeFinalDecisionDoc(CaseData caseData, CaseData oldData) {
+        List<ListValue<CaseworkerCICDocument>> wholeFinalDecisionDocList = DocumentListUtil.getAllFinalDecisionDocuments(oldData);
+
+        if (wholeFinalDecisionDocList.size() > caseData.getCicCase().getFinalDecisionDocumentList().size()) {
+            for (ListValue<CaseworkerCICDocument> cicDocumentListValue : wholeFinalDecisionDocList) {
+                checkFinalDecision(caseData, cicDocumentListValue);
+            }
+        }
+    }
+
+    public static void removeDecisionDoc(CaseData caseData, CaseData oldData) {
+        List<ListValue<CaseworkerCICDocument>> wholeDecisionDocList = DocumentListUtil.getAllDecisionDocuments(oldData);
+
+        if (wholeDecisionDocList.size() > caseData.getCicCase().getDecisionDocumentList().size()) {
+            for (ListValue<CaseworkerCICDocument> doc : wholeDecisionDocList) {
+                if (!caseData.getCicCase().getDecisionDocumentList().contains(doc)
+                    && doc.getValue().getDocumentLink().equals(caseData.getCaseIssueDecision().getIssueDecisionDraft())) {
+                    caseData.getCaseIssueDecision().setIssueDecisionDraft(null);
+                } else if (doc.getValue().getDocumentLink()
+                    .equals(caseData.getCaseIssueDecision().getDecisionDocument().getDocumentLink())) {
+                    caseData.getCaseIssueDecision().setDecisionDocument(EMPTY_DOCUMENT);
+                }
+                addToRemovedDocuments(caseData.getCicCase(), doc.getValue());
+            }
+        }
+    }
+
+    public static void checkFinalDecision(CaseData caseData, ListValue<CaseworkerCICDocument> cicDocumentListValue) {
+        if (!caseData.getCicCase().getFinalDecisionDocumentList().contains(cicDocumentListValue)) {
+            if (cicDocumentListValue.getValue().getDocumentLink()
+                .equals(caseData.getCaseIssueFinalDecision().getFinalDecisionDraft())) {
+                caseData.getCaseIssueFinalDecision().setFinalDecisionDraft(null);
+            } else if (cicDocumentListValue.getValue().getDocumentLink()
+                .equals(caseData.getCaseIssueFinalDecision().getDocument().getDocumentLink())) {
+                caseData.getCaseIssueFinalDecision().setDocument(EMPTY_DOCUMENT);
+            }
+            addToRemovedDocuments(caseData.getCicCase(), cicDocumentListValue.getValue());
+        }
+
     }
 
 }
