+++ src/main/java/uk/gov/hmcts/sptribs/caseworker/util/OrderDocumentListUtil.java
@@ -12,6 +12,9 @@ import uk.gov.hmcts.sptribs.document.model.DocumentType;
 import java.util.ArrayList;
 import java.util.List;
 
+import static uk.gov.hmcts.sptribs.caseworker.util.DocumentManagementUtil.EMPTY_DOCUMENT;
+import static uk.gov.hmcts.sptribs.caseworker.util.DocumentManagementUtil.addToRemovedDocuments;
+
 public final class OrderDocumentListUtil {
 
     private OrderDocumentListUtil() {
@@ -53,4 +56,41 @@ public final class OrderDocumentListUtil {
         }
         return orderUploadList;
     }
+
+
+    public static CicCase removeOrderDoc(CicCase cicCase, CicCase oldCicCase) {
+        List<ListValue<CaseworkerCICDocument>> wholeOrderDocList = DocumentListUtil.getAllOrderDocuments(oldCicCase);
+
+        if (wholeOrderDocList.size() > cicCase.getOrderDocumentList().size()) {
+            for (ListValue<CaseworkerCICDocument> cicDocumentListValue : wholeOrderDocList) {
+                if (!cicCase.getOrderDocumentList().contains(cicDocumentListValue)) {
+                    for (ListValue<Order> orderListValue : cicCase.getOrderList()) {
+                        if (null != orderListValue.getValue().getDraftOrder()
+                            && cicDocumentListValue.getValue().getDocumentLink()
+                            .equals(orderListValue.getValue().getDraftOrder().getTemplateGeneratedDocument())) {
+                            orderListValue.getValue().getDraftOrder().setTemplateGeneratedDocument(null);
+                        } else {
+                            manageUploadedFiles(orderListValue, cicDocumentListValue);
+                        }
+                    }
+                    addToRemovedDocuments(cicCase, cicDocumentListValue.getValue());
+                }
+            }
+        }
+        return cicCase;
+    }
+
+    public static void manageUploadedFiles(ListValue<Order> orderListValue, ListValue<CaseworkerCICDocument> cicDocumentListValue) {
+        if (!CollectionUtils.isEmpty(orderListValue.getValue().getUploadedFile())) {
+            for (int i = 0; i < orderListValue.getValue().getUploadedFile().size(); i++) {
+                ListValue<CICDocument> file = orderListValue.getValue().getUploadedFile().get(i);
+                if (null != file.getValue().getDocumentLink() && file.getValue().getDocumentLink()
+                    .equals(cicDocumentListValue.getValue().getDocumentLink())) {
+                    orderListValue.getValue().getUploadedFile().get(i).setValue(EMPTY_DOCUMENT);
+
+                }
+            }
+        }
+    }
+
 }
