+++ src/test/java/uk/gov/hmcts/sptribs/caseworker/util/DocumentListUtilTest.java
@@ -26,10 +26,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.caseData;
+import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.get2Document;
+import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.getDocument;
 
 @ExtendWith(MockitoExtension.class)
 public class DocumentListUtilTest {
-
 
     @Test
     void shouldGenerateDocList() {
@@ -210,5 +212,58 @@ public class DocumentListUtilTest {
         assertThat(result).isNotNull();
     }
 
+    @Test
+    void shouldSuccessfullyRemoveDocs() {
+        //Given
+        final CaseData caseData = caseData();
+        CICDocument doc = CICDocument.builder()
+            .documentLink(Document.builder().url("url1").binaryUrl("url1").filename("name1").build()).build();
+        caseData.setCaseIssueFinalDecision(CaseIssueFinalDecision.builder().document(doc).build());
+        caseData.setCaseIssueDecision(CaseIssueDecision.builder().decisionDocument(doc).build());
+
+        CICDocument document = CICDocument.builder()
+            .documentLink(Document.builder().url("url1").binaryUrl("url1").filename("name1").build()).build();
+        ListValue<CICDocument> cicDocumentListValue = new ListValue<>();
+        cicDocumentListValue.setValue(document);
+        Order order = Order.builder().uploadedFile(List.of(cicDocumentListValue)).build();
+        ListValue<Order> orderListValue = new ListValue<>();
+        orderListValue.setValue(order);
+
+
+        CicCase cicCase = CicCase.builder()
+            .orderDocumentList(getDocument())
+            .orderList(List.of(orderListValue))
+            .decisionDocumentList(new ArrayList<>())
+            .finalDecisionDocumentList(new ArrayList<>())
+            .applicantDocumentsUploaded(getDocument())
+            .build();
+        caseData.setCicCase(cicCase);
+        DocumentManagement documentManagement = DocumentManagement.builder().caseworkerCICDocument(getDocument()).build();
+        caseData.setDocManagement(documentManagement);
+        final CaseData oldData = caseData();
+
+        CICDocument docOld = CICDocument.builder()
+            .documentLink(Document.builder().url("url1").binaryUrl("url1").filename("name1").build()).build();
+        oldData.setCaseIssueFinalDecision(CaseIssueFinalDecision.builder().document(docOld).build());
+        oldData.setCaseIssueDecision(CaseIssueDecision.builder().decisionDocument(docOld).build());
+
+        DocumentManagement documentManagementOld = DocumentManagement.builder().caseworkerCICDocument(get2Document()).build();
+        oldData.setDocManagement(documentManagementOld);
+        CicCase cicCaseOld = CicCase.builder()
+            .decisionDocumentList(get2Document())
+            .finalDecisionDocumentList(get2Document())
+            .applicantDocumentsUploaded(get2Document())
+            .reinstateDocuments(get2Document())
+            .build();
+        oldData.setCicCase(cicCaseOld);
+        oldData.setListing(Listing.builder().summary(HearingSummary.builder().recFile(get2Document()).build()).build());
+        oldData.setCloseCase(CloseCase.builder().documents(get2Document()).build());
+
+        //When
+        CaseData result = DocumentListUtil.removeEvaluatedListDoc(caseData, oldData);
+
+        //Then
+        assertThat(result).isNotNull();
+    }
 
 }
